refactor: closedStudyItem 폴더 및 네이밍 변경
