+++ src/app/profile/closedStudyList/ClosedStudyItem.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { IoEllipsisHorizontal, IoEllipsisVertical } from 'react-icons/io5';
 import moment from 'moment';
 import NumberUtils from '@src/app/shared/utils/number';
-import { AppliedStudyInterface } from '../../interface/interface';
+import { AppliedStudyInterface } from '../interface/interface';
 import './index.scss';
 
-interface ClosedAppliedStudyItemProps {
-	closedAppliedStudyItem: AppliedStudyInterface;
+interface ClosedStudyItemProps {
+	closedStudyItem: AppliedStudyInterface;
 }
 
-const ClosedAppliedStudyItem = ({ closedAppliedStudyItem }: ClosedAppliedStudyItemProps): JSX.Element => {
+const ClosedStudyItem = ({ closedStudyItem }: ClosedStudyItemProps): JSX.Element => {
 	const onClickMenu = () => {
 		console.log('onClickMenu');
 	};
@@ -17,23 +17,23 @@ const ClosedAppliedStudyItem = ({ closedAppliedStudyItem }: ClosedAppliedStudyIt
 	return (
 		<div className="closed-applied-study-item-container">
 			<div className="closed-applied-study-item-column-container">
-				<div className="closed-applied-study-item-title">{closedAppliedStudyItem.title}</div>
+				<div className="closed-applied-study-item-title">{closedStudyItem.title}</div>
 				<div className="closed-applied-study-item-row-container">
 					<div className="closed-applied-study-item-subtext">
-						{moment(closedAppliedStudyItem.date).format('YY.MM.DD HH:MM')}
+						{moment(closedStudyItem.date).format('YY.MM.DD HH:MM')}
 					</div>
 					<div className="closed-applied-study-item-divider-dot">·</div>
 					<div className="closed-applied-study-item-subtext">
-						조회 {NumberUtils.toFormattedCount(closedAppliedStudyItem.hits)}
+						조회 {NumberUtils.toFormattedCount(closedStudyItem.hits)}
 					</div>
 					<div className="closed-applied-study-item-divider-dot">·</div>
 					<div className="closed-applied-study-item-subtext">
-						북마크 {NumberUtils.toFormattedCount(closedAppliedStudyItem.bookmarks)}
+						북마크 {NumberUtils.toFormattedCount(closedStudyItem.bookmarks)}
 					</div>
 				</div>
 			</div>
 			<div className="closed-applied-study-item-row-container">
-				{closedAppliedStudyItem.status === 'accepted' && (
+				{closedStudyItem.status === 'accepted' && (
 					<div className="closed-applied-study-item-status">참여완료</div>
 				)}
 				<button type="button" onClick={onClickMenu}>
@@ -44,4 +44,4 @@ const ClosedAppliedStudyItem = ({ closedAppliedStudyItem }: ClosedAppliedStudyIt
 	);
 };
 
-export default ClosedAppliedStudyItem;
+export default ClosedStudyItem;
