+++ src/app/profile/closedStudyList/index.scss
