+++ src/app/profile/openedAppliedStudyList/index.scss
